Rename toasted.ts to SetupToasted.ts to match default export
